+++ src/app/context/CurrencyContext.tsx
@@ -1,10 +1,13 @@
 import { useState, useEffect, createContext, useContext } from "react";
+// https://www.npmjs.com/package/currency.js 0 dependencies, 135,987k weekly downloads
+import currencyJs from "currency.js";
+// https://github.com/MikeMcl/decimal.js-light 0 dependencies, 826,917 weekly downloads
+import { Decimal } from "decimal.js-light";
+import axios from "axios";
+
 import { useAuth } from "./AuthContext";
 import api from "~/common/lib/api";
-import {
-  satoshisToFiat,
-  SupportedCurrencies,
-} from "../utils/currencyConversion";
+import { SupportedCurrencies, SupportedExchanges } from "~/types";
 
 interface CurrencyContextType {
   balances: {
@@ -17,6 +20,8 @@ interface CurrencyContextType {
 }
 
 const CurrencyContext = createContext({} as CurrencyContextType);
+
+const numSatsInBtc = 100_000_000;
 
 export function CurrencyProvider({ children }: { children: React.ReactNode }) {
   const currencies = [
@@ -194,7 +199,57 @@ export function CurrencyProvider({ children }: { children: React.ReactNode }) {
     fiatBalance: string;
   }>({ satsBalance: "", fiatBalance: "" });
   const [currency, setCurrency] = useState<SupportedCurrencies>("USD");
+  const [exchange, setExchange] = useState<SupportedExchanges>("Coindesk");
   const auth = useAuth();
+
+  const bitcoinToFiat = async (
+    amountInBtc: number | string,
+    convertTo: SupportedCurrencies
+  ) => {
+    const btc = new Decimal(amountInBtc);
+    const rate = await getFiatBtcRate(convertTo);
+    return btc.mul(rate).toNumber();
+  };
+
+  const getFiatBtcRate = async (
+    currency: SupportedCurrencies
+  ): Promise<string> => {
+    let response;
+
+    if (exchange === "Yad.io") {
+      response = await axios.get(
+        `https://api.yadio.io/exrates/${currency.toLowerCase()}`
+      );
+      const data = await response?.data;
+      return currencyJs(data.BTC, {
+        separator: "",
+        symbol: "",
+      }).format();
+    }
+
+    response = await axios.get(
+      `https://api.coindesk.com/v1/bpi/currentprice/${currency.toLowerCase()}.json`
+    );
+    const data = await response?.data;
+    return currencyJs(data.bpi[currency].rate, {
+      separator: "",
+      symbol: "",
+    }).format();
+  };
+
+  const satoshisToBitcoin = (amountInSatoshis: number | string) => {
+    const sats = new Decimal(amountInSatoshis);
+    return sats.div(numSatsInBtc).toNumber();
+  };
+
+  const satoshisToFiat = async (
+    amountInSats: number | string,
+    convertTo: SupportedCurrencies
+  ) => {
+    const btc = satoshisToBitcoin(amountInSats);
+    const fiat = await bitcoinToFiat(btc, convertTo);
+    return fiat;
+  };
 
   const getBalances = async (balance: number) => {
     const fiatValue = await satoshisToFiat(balance, currency);
@@ -221,7 +276,10 @@ export function CurrencyProvider({ children }: { children: React.ReactNode }) {
     setCurrency(currency);
 
   useEffect(() => {
-    api.getSettings().then((settings) => setCurrency(settings.currency));
+    api.getSettings().then((settings) => {
+      setCurrency(settings.currency);
+      setExchange(settings.exchange);
+    });
   }, []);
 
   useEffect(() => {
