+++ src/app/screens/Settings.tsx
@@ -10,10 +10,13 @@ import { useState, useEffect } from "react";
 import { toast } from "react-toastify";
 import { getTheme } from "~/app/utils";
 import api from "~/common/lib/api";
-import { SettingsStorage } from "~/types";
+import {
+  SettingsStorage,
+  SupportedCurrencies,
+  SupportedExchanges,
+} from "~/types";
 
 import { useCurreny } from "../context/CurrencyContext";
-import { SupportedCurrencies } from "../utils/currencyConversion";
 
 function Settings() {
   const [loading, setLoading] = useState(true);
@@ -26,7 +29,7 @@ function Settings() {
     locale: "",
     theme: "system",
     currency: "USD",
-    exchange: "",
+    exchange: "Coindesk",
   });
   const [cameraPermissionsGranted, setCameraPermissionsGranted] =
     useState(false);
@@ -167,11 +170,12 @@ function Settings() {
                 value={settings.exchange}
                 onChange={async (ev) => {
                   await saveSetting({
-                    exchange: ev.target.value,
+                    exchange: ev.target.value as SupportedExchanges,
                   });
                 }}
               >
-                <option value="coindesk">Coindesk</option>
+                <option value="Coindesk">Coindesk</option>
+                <option value="Yad.io">Yad.io</option>
               </Select>
             </div>
           )}
