+++ src/extension/background-script/state.ts
@@ -2,10 +2,15 @@ import merge from "lodash/merge";
 import pick from "lodash/pick";
 import browser from "webextension-polyfill";
 import createState from "zustand";
-import { SupportedCurrencies } from "~/app/utils/currencyConversion";
 import { decryptData } from "~/common/lib/crypto";
 import i18n from "~/i18n/i18nConfig";
-import type { Account, Accounts, SettingsStorage } from "~/types";
+import type {
+  Account,
+  Accounts,
+  SettingsStorage,
+  SupportedCurrencies,
+  SupportedExchanges,
+} from "~/types";
 
 import connectors from "./connectors";
 import type Connector from "./connectors/connector.interface";
@@ -39,7 +44,7 @@ export const DEFAULT_SETTINGS = {
   locale: i18n.resolvedLanguage,
   theme: "system",
   currency: "USD" as SupportedCurrencies,
-  exchange: "",
+  exchange: "Coindesk" as SupportedExchanges,
 };
 
 // these keys get synced from the state to the browser storage
